pyproject.toml
[project]
name = "mdbook"
version = "3.0.0"
description = "MD Book Tools - A DI/SOA markdown book reader and writer with MCP server support"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
keywords = ["markdown", "book", "reader", "writer", "terminal", "mdbook", "gitbook", "leanpub", "bookdown", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "markdown>=3.4.0",
    "pymdown-extensions>=9.0.0",
    "pygments>=2.14.0",
    "pyyaml>=6.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
toml = ["tomli>=2.0.0;python_version<'3.11'"]

[project.scripts]
mdbook = "mdbook.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["mdbook*"]
